tests: Use babel-plugin-tester to write tests
