+++ package.json
@@ -10,7 +10,9 @@
   "engines": {
     "node": ">= 6.0.0"
   },
-  "files": ["lib"],
+  "files": [
+    "lib"
+  ],
   "author": {
     "name": "Tommy Leunen",
     "email": "tommy.leunen@gmail.com",
@@ -42,6 +44,7 @@
     "babel-core": "^6.26.0",
     "babel-jest": "^22.4.1",
     "babel-plugin-syntax-dynamic-import": "^6.18.0",
+    "babel-plugin-tester": "^5.0.0",
     "babel-plugin-transform-es2015-modules-commonjs": "^6.26.0",
     "babel-plugin-transform-object-rest-spread": "^6.26.0",
     "babel-preset-env": "^1.6.0",
@@ -57,7 +60,7 @@
     "standard-version": "^4.2.0"
   },
   "scripts": {
-    "lint": "eslint src test",
+    "lint": "eslint src test --ignore-pattern test/fixtures/**/*.js",
     "compile": "babel src --out-dir lib",
     "pretest": "npm run lint",
     "test": "jest",
@@ -68,11 +71,17 @@
     "precommit": "lint-staged"
   },
   "lint-staged": {
-    "*.{json,md}": ["prettier-eslint --write", "git add"],
-    "*.{js}": ["prettier-eslint --write", "eslint --fix", "git add"]
+    "*.js": [
+      "prettier-eslint --write",
+      "eslint --fix --ignore-pattern test/fixtures/**/*.js",
+      "git add"
+    ]
   },
   "jest": {
     "testRegex": "/test/.*\\.test\\.js$",
-    "collectCoverageFrom": ["src/**/*.js", "!src/log.js"]
+    "collectCoverageFrom": [
+      "src/**/*.js",
+      "!src/log.js"
+    ]
   }
 }
