+++ test/fixtures/dynamic-import/.babelrc
@@ -1,0 +1,14 @@
+{
+  "plugins": [
+    "syntax-dynamic-import",
+    ["../../../src/index.js", {
+      "root": "./test/testproject/src",
+      "alias": {
+        "test": "./test/testproject/test"
+      },
+      "transformFunctions": [
+        "customMethod.something"
+      ]
+    }]
+  ]
+}
