+++ test/fixtures/imports-from-other-plugins/.babelrc
@@ -1,0 +1,9 @@
+{
+  "plugins": [
+    "./fakePlugin.js",
+    "syntax-dynamic-import",
+    ["../../../src/index.js", {
+      "root": "./test/testproject/src"
+    }]
+  ]
+}
