+++ test/fixtures/imports-from-other-plugins/code.js
@@ -1,0 +1,1 @@
+bootest('components/Header/SubHeader');
