+++ test/fixtures/imports-from-other-plugins/fakePlugin.js
@@ -1,0 +1,9 @@
+module.exports = function fakePlugin({ types }) {
+  return {
+    visitor: {
+      Identifier(path) {
+        path.replaceWith(types.Import());
+      },
+    },
+  };
+};
