+++ test/fixtures/imports-from-other-plugins/output.js
@@ -1,0 +1,1 @@
+import('./test/testproject/src/components/Header/SubHeader');
